Refactor dom utilities into a shared location
