+++ src/parse.ts
@@ -2,23 +2,21 @@ import * as csstree from 'css-tree';
 import { nanoid } from 'nanoid/non-secure';
 
 import {
+  getCSSPropertyValue,
+  type PseudoElement,
+  type Selector,
+} from './dom.js';
+import {
   type DeclarationWithValue,
   generateCSS,
   getAST,
   getDeclarationValue,
-  SHIFTED_PROPERTIES,
   type StyleData,
 } from './utils.js';
-import { type PseudoElement, validatedForPositioning } from './validate.js';
+import { validatedForPositioning } from './validate.js';
 
 interface AtRuleRaw extends csstree.Atrule {
   prelude: csstree.Raw | null;
-}
-
-export interface Selector {
-  selector: string;
-  elementPart: string;
-  pseudoElementPart?: string;
 }
 
 // `key` is the `anchor-name` value
@@ -472,10 +470,6 @@ function getPositionFallbackRules(node: csstree.Atrule) {
   return {};
 }
 
-export function getCSSPropertyValue(el: HTMLElement, prop: string) {
-  return getComputedStyle(el).getPropertyValue(prop).trim();
-}
-
 async function getAnchorEl(
   targetEl: HTMLElement | null,
   anchorObj: AnchorFunction,
@@ -486,7 +480,7 @@ async function getAnchorEl(
     const anchorAttr = targetEl.getAttribute('anchor');
     const positionAnchorProperty = getCSSPropertyValue(
       targetEl,
-      SHIFTED_PROPERTIES['position-anchor'],
+      'position-anchor',
     );
 
     if (positionAnchorProperty) {
