+++ src/utils.ts
@@ -34,4 +34,5 @@ export interface StyleData {
 export const SHIFTED_PROPERTIES: Record<string, string> = {
   'position-anchor': `--position-anchor-${nanoid(12)}`,
   'anchor-scope': `--anchor-scope-${nanoid(12)}`,
+  'anchor-name': `--anchor-name-${nanoid(12)}`,
 };
