+++ src/validate.ts
@@ -1,152 +1,12 @@
-import { platform, type VirtualElement } from '@floating-ui/dom';
-import { nanoid } from 'nanoid/non-secure';
+import { platform } from '@floating-ui/dom';
 
 import {
-  AnchorScopeValue,
   getCSSPropertyValue,
+  getElementsBySelector,
+  hasStyle,
   type Selector,
-} from './parse.js';
-import { SHIFTED_PROPERTIES } from './utils.js';
-
-export interface PseudoElement extends VirtualElement {
-  fakePseudoElement: HTMLElement;
-  computedStyle: CSSStyleDeclaration;
-  removeFakePseudoElement(): void;
-}
-
-function findFirstScrollingElement(element: HTMLElement) {
-  let currentElement: HTMLElement | null = element;
-
-  while (currentElement) {
-    if (hasStyle(currentElement, 'overflow', 'scroll')) {
-      return currentElement;
-    }
-
-    currentElement = currentElement.parentElement;
-  }
-
-  return currentElement;
-}
-
-/**
-  Like `document.querySelectorAll`, but if the selector has a pseudo-element
-  it will return a wrapper for the rest of the polyfill to use.
-*/
-function getAnchorsBySelectors(selectors: Selector[]) {
-  const result: (HTMLElement | PseudoElement)[] = [];
-
-  for (const { selector, elementPart, pseudoElementPart } of selectors) {
-    const isBefore = pseudoElementPart === '::before';
-    const isAfter = pseudoElementPart === '::after';
-
-    // Current we only support `::before` and `::after` pseudo-elements.
-    if (pseudoElementPart && !(isBefore || isAfter)) continue;
-
-    const elements = Array.from(
-      document.querySelectorAll<HTMLElement>(elementPart),
-    );
-
-    if (!pseudoElementPart) {
-      result.push(...elements);
-      continue;
-    }
-
-    for (const element of elements) {
-      // Floating UI needs `Element.getBoundingClientRect` to calculate the position for the anchored element,
-      // since there isn't a way to get it for pseudo-elements;
-      // we create a temporary "fake pseudo-element" that we use as reference.
-      const computedStyle = getComputedStyle(element, pseudoElementPart);
-      const fakePseudoElement = document.createElement('div');
-      const sheet = document.createElement('style');
-
-      fakePseudoElement.id = `fake-pseudo-element-${nanoid()}`;
-
-      // Copy styles from pseudo-element to the "fake pseudo-element", `.cssText` does not work on Firefox.
-      for (const property of Array.from(computedStyle)) {
-        const value = computedStyle.getPropertyValue(property);
-        fakePseudoElement.style.setProperty(property, value);
-      }
-
-      // For the `content` property, since normal elements don't have it,
-      // we add the content to a pseudo-element of the "fake pseudo-element".
-      sheet.textContent += `#${fakePseudoElement.id}${pseudoElementPart} { content: ${computedStyle.content}; }`;
-      // Hide the pseudo-element while the "fake pseudo-element" is visible.
-      sheet.textContent += `${selector} { display: none !important; }`;
-
-      document.head.append(sheet);
-
-      if (isBefore) {
-        element.insertAdjacentElement('afterbegin', fakePseudoElement);
-      }
-
-      if (isAfter) {
-        element.insertAdjacentElement('beforeend', fakePseudoElement);
-      }
-
-      const boundingClientRect = fakePseudoElement.getBoundingClientRect();
-
-      const { scrollY: startingScrollY, scrollX: startingScrollX } = globalThis;
-      let firstScrollingElement: {
-        scrollTop: number;
-        scrollLeft: number;
-      } | null = findFirstScrollingElement(element);
-
-      // Avoid doubled scroll
-      if (firstScrollingElement === document.documentElement) {
-        firstScrollingElement = null;
-      }
-
-      firstScrollingElement ??= { scrollTop: 0, scrollLeft: 0 };
-
-      const { scrollTop: startingScrollTop, scrollLeft: startingScrollLeft } =
-        firstScrollingElement;
-
-      result.push({
-        // Passed to `isAcceptableAnchorElement`.
-        fakePseudoElement,
-        // For testing.
-        computedStyle,
-
-        // For https://floating-ui.com/docs/autoupdate#ancestorscroll to work on `VirtualElement`s.
-        contextElement: element,
-
-        // For `validatedForPositioning` to "undo" the "fake pseudo-element" after it's been used.
-        removeFakePseudoElement() {
-          fakePseudoElement.remove();
-          sheet.remove();
-        },
-
-        // https://floating-ui.com/docs/virtual-elements.
-        getBoundingClientRect() {
-          const { scrollY, scrollX } = globalThis;
-          const { scrollTop, scrollLeft } = firstScrollingElement;
-
-          return DOMRect.fromRect({
-            y:
-              boundingClientRect.y +
-              (startingScrollY - scrollY) +
-              (startingScrollTop - scrollTop),
-            x:
-              boundingClientRect.x +
-              (startingScrollX - scrollX) +
-              (startingScrollLeft - scrollLeft),
-
-            width: boundingClientRect.width,
-            height: boundingClientRect.height,
-          });
-        },
-      });
-    }
-  }
-
-  return result;
-}
-
-// Given an element and CSS style property,
-// checks if the CSS property equals a certain value
-function hasStyle(element: HTMLElement, cssProperty: string, value: string) {
-  return getCSSPropertyValue(element, cssProperty) === value;
-}
+} from './dom.js';
+import { AnchorScopeValue } from './parse.js';
 
 // Given a target element's containing block (CB) and an anchor element,
 // determines if the anchor element is a descendant of the target CB.
@@ -330,10 +190,7 @@ function hasScope(
   // Just because the element matches a rule that sets the scope we're looking for, does not mean
   // that that rule is actually selected in the cascade. We read our `--anchor-scope` custom
   // property to confirm which rule is actually applied.
-  const computedScope = getCSSPropertyValue(
-    element,
-    SHIFTED_PROPERTIES['anchor-scope'],
-  );
+  const computedScope = getCSSPropertyValue(element, 'anchor-scope');
   return computedScope === anchorName || computedScope === AnchorScopeValue.All;
 }
 
@@ -359,7 +216,10 @@ export async function validatedForPositioning(
     return null;
   }
 
-  const anchorElements = getAnchorsBySelectors(anchorSelectors);
+  const anchorElements = anchorSelectors
+    // Any element that matches a selector that sets the specified `anchor-name` could be a
+    // potential match.
+    .flatMap(getElementsBySelector);
 
   // TODO: handle anchor-scope for pseudo-elements.
   const scopeSelector = scopeSelectors.map((s) => s.selector).join(',') || null;
